Predict the name of the book
This Code aims to produce classification predictions and compare them; analyze the pros and
cons of algorithms and generate and communicate the insights and discuss the implementation
steps of applying (preprocessing the data, transformation and evaluation,etc) strategies on three
of multiclass classification algorithms (k-Nearest Neighbor, Support Vector Machine and
Decision Tree algorithm). I will introduce the detailed steps of implementation of this strategy
and discuss the achieved results. I used (five different samples of Gutenberg digital books) for
implementation.