text clustering
This report aims to produce is to produce similar clusters and compare them; analyze the pros
and cons of algorithms, generate and communicate the insights and discuss the implementation
steps of applying (preprocessing the data, transformation and evaluation,etc) strategies on three
of Clustering algorithms (K-means, EM, Hierarchical clustering algorithm). We will introduce
the detailed steps of implementation of this strategy and discuss the achieved results. We used
(five different samples of Gutenberg digital books) for implementation.